requirements.txt
biopython==1.79
docopt==0.6.2
numpy==1.21.2
